+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/domene/VedtaksperiodeMedBegrunnelser.kt
@@ -4,9 +4,10 @@ import com.fasterxml.jackson.annotation.JsonIgnore
 import no.nav.familie.ba.sak.common.BaseEntitet
 import no.nav.familie.ba.sak.common.Feil
 import no.nav.familie.ba.sak.common.NullablePeriode
+import no.nav.familie.ba.sak.common.inneværendeMåned
+import no.nav.familie.ba.sak.common.toYearMonth
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.hentAndelerForSegment
-import no.nav.familie.ba.sak.kjerne.beregning.domene.hentLøpendeAndelForVedtaksperiode
 import no.nav.familie.ba.sak.kjerne.brev.domene.SanityBegrunnelse
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlag
 import no.nav.familie.ba.sak.kjerne.vedtak.Vedtak
@@ -16,7 +17,9 @@ import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.Vedtaksperiodetype
 import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.lagUtbetalingsperiodeDetaljer
 import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.utledSegmenter
 import no.nav.familie.ba.sak.sikkerhet.RollestyringMotDatabase
+import no.nav.fpsak.tidsserie.LocalDateSegment
 import org.hibernate.annotations.SortComparator
+import org.slf4j.LoggerFactory
 import java.time.LocalDate
 import javax.persistence.CascadeType
 import javax.persistence.Column
@@ -124,15 +127,31 @@ data class VedtaksperiodeMedBegrunnelser(
                     it.endretUtbetalingAndeler.isEmpty()
                 }
             }
-            val vertikaltSegmentForVedtaksperiode =
-                if (this.type == Vedtaksperiodetype.FORTSATT_INNVILGET)
-                    andelerForVedtaksperiodetype.hentLøpendeAndelForVedtaksperiode()
-                else hentVertikaltSegmentForVedtaksperiode(andelerForVedtaksperiodetype)
+            if (andelerForVedtaksperiodetype.isEmpty()) emptyList()
+            else {
+                val vertikaltSegmentForVedtaksperiode =
+                    if (this.type == Vedtaksperiodetype.FORTSATT_INNVILGET)
+                        hentLøpendeAndelForVedtaksperiode(andelerForVedtaksperiodetype)
+                    else hentVertikaltSegmentForVedtaksperiode(andelerForVedtaksperiodetype)
 
-            val andelerForSegment =
-                andelerForVedtaksperiodetype.hentAndelerForSegment(vertikaltSegmentForVedtaksperiode)
+                if (vertikaltSegmentForVedtaksperiode == null) {
+                    LoggerFactory.getLogger(VedtaksperiodeMedBegrunnelser::class.java)
+                        .error("Finner ikke segment for vedtaksperiode $this. Se securelogger for mer informasjon.")
+                    LoggerFactory.getLogger("secureLogger")
+                        .info(
+                            "Finner ikke segment for vedtaksperiode $this.\n " +
+                                "Alle andeler=$andelerTilkjentYtelse.\n" +
+                                "AndelerForVedtaksperiode=$andelerForVedtaksperiodetype."
+                        )
 
-            andelerForSegment.lagUtbetalingsperiodeDetaljer(personopplysningGrunnlag)
+                    emptyList()
+                } else {
+                    val andelerForSegment =
+                        andelerForVedtaksperiodetype.hentAndelerForSegment(vertikaltSegmentForVedtaksperiode)
+
+                    andelerForSegment.lagUtbetalingsperiodeDetaljer(personopplysningGrunnlag)
+                }
+            }
         } else {
             emptyList()
         }
@@ -148,6 +167,13 @@ data class VedtaksperiodeMedBegrunnelser(
     companion object {
         val comparator = BegrunnelseComparator()
     }
+}
+
+private fun hentLøpendeAndelForVedtaksperiode(andelerTilkjentYtelse: List<AndelTilkjentYtelse>): LocalDateSegment<Int> {
+    val sorterteSegmenter = andelerTilkjentYtelse.utledSegmenter().sortedBy { it.fom }
+    return sorterteSegmenter.lastOrNull { it.fom.toYearMonth() <= inneværendeMåned() }
+        ?: sorterteSegmenter.firstOrNull()
+        ?: throw Feil("Finner ikke gjeldende segment ved fortsatt innvilget")
 }
 
 class BegrunnelseComparator : Comparator<Vedtaksbegrunnelse> {
